fix(slicing): accumulate seam layer z in f64
Upstream SeamPlacer stores each layer's slice_z from a double print_z
accumulation; the Rust port accumulated in f32 and drifted a few ulps
per layer. Accumulate in f64 and cast per layer: aligned-spline inputs
now match the oracle dump for 39/42 strings (was 0/42), and 1um seam
divergences drop from 2594 to 2453 of 5243 placements.

+++ crates/ares-core/src/project_slice/seam_placement.rs
@@ -80,13 +80,15 @@ fn apply_objects(
         if !aligned[object_index] {
             continue;
         }
+        // Upstream `slice_z` is a double accumulated by the slicing loop; accumulate
+        // in f64 and cast per layer so candidate Z matches OrcaSlicer bit-for-bit.
         let layer_zs = traversal.objects[object_index]
             .records
             .iter()
-            .scan(0.0_f32, |print_z, record| {
-                let height = record.as_ref().map_or(0.0, |record| record.layer_height) as f32;
+            .scan(0.0_f64, |print_z, record| {
+                let height = record.as_ref().map_or(0.0, |record| record.layer_height);
                 *print_z += height;
-                Some(*print_z - 0.5 * height)
+                Some((*print_z - 0.5 * height) as f32)
             })
             .collect::<Vec<_>>();
         let mut plans = alignment::prepare(
